Code update: counter counts the galaxies
This code:

count_test (as before) - uses circle data, with background removed, locates itself a centre (since no peak in middle like the galaxies), then creates a circle of fixed(or variable) r. For fixed case - works if r>30, for variable r, must increase in steps of 5 to find one object currently.

Galaxy count - changed to a 5sigma limit on galaxy data, this is used to determine the minimum brightness of a galaxy, and also, set all masked data to this value so it counts everything above this.. takes a while to run..
It uses circular shape and fixed aperture size, but calculates a local background for each star when it determines the magnitude.

Next: Test on small section to show that it works and also for different backgrounds (sigma) and a crude quick method can be ignoring local background, subtract the global background from each star to show the difference here.. then think about varying r and accounting for non circular galaxies